Restructured the client React components folder - Added absolute path import support - Updated all component imports
