+++ project-wesley-wei-client/src/components/pages/landing/Landing.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Compass from '../Compass/Compass';
-import Navbar from '../Navigation/Navbar';
-import Ocean from '../3DModels/Ocean';
+import Compass from 'components/common/compass/Compass';
+import Navbar from 'components/layout/navigation/Navbar';
+import Ocean from 'components/common/3DModels/Ocean';
 import { Canvas } from '@react-three/fiber';
 import { Sky } from '@react-three/drei';
 import './styles/Landing.css';
